+++ src/world/light.hpp
@@ -23,7 +23,7 @@ enum class ps2_blend_mode : int8 { add = 0, multiply = 1, blend = 2 };
 
 struct light {
    std::string name;
-   int16 layer = 0;
+   int8 layer = 0;
    bool hidden = false;
 
    quaternion rotation = {1.0f, 0.0f, 0.0f, 0.0f};
